FINAL!
Moved update function's steps into settings.php - removed update.php
entirely
Final cleanup of settings.php for update purposes

The Update Application button FINALLY works! When clicked, it will pull
down the entire branch of your choosing and overwrite all of the files.
(Note for future: The "overwrite" does not delete files that are no
longer present. Therefore, if any files are deleted in the future, an
unlink will NEED to be included for the cleanup.)