Fix `sold_date` error
After connecting the sold_date field to the PDO call, the call would error when a date was not entered.

This was caused by the improper passing of an empty string to the SQL date field.

To fix this, an if statement was added to check for an empty string and convert it to a null if found.