Fix bug where save deletes branch
When the Save button was pressed on the Database tab, the selected branch would be deleted from the INI file.

This happened because a blank string value was passed to the `update_config()` function, and when it was tested using the `isset()` function, it returned true because the variable was set.

This was fixed by using `!empty()` instead, which would register as false and move onto the `elseif`.

Updated the `$commit` test as well.